build: rewrite source module
Source files are now managed by attaching them to explicit groups.
Groups cannot be nested.  This allows the tree of an IDE project to
mirror the filesystem structure except for the paths of top-level
groups, which can then be located outside the build tree.

Source files now have methods for working with the atoms they have.