build: add better support for built sources
Build modules are now scanned in two phases.  In the first phase, they have
the opportunity to add built sources to the project, as well as targets that
build the sources.  In the second phase, they use the project to build
targets.  A new pseudotarget 'built-sources' will build all built sources.

The source module was updated to make it easier to add sources to existing
groups.