build: major revision of build system
Build environments are now divided into separate objects: executables
and projects have info dictionaries, build settings are in one object,
and there is still the build environment.  This new division removes
many of the warts from the existing code.

The concept of modules is added, allowing more sophisticated conditional
compilation.  For example, it is now possible to disable the C++ parts
of the sglib library.  The linker target has been updated to only use
g++ when necessary.