Fix resource loading on Mac OS X
This was an aggravating bug to quash.  It turns out that Apple's GCC 4.2
doesn't emit an error when processing a packed enum attribute, but
neither does it work correctly.  Changing the attribute syntax so the
attribute goes after the final brace seems to make it work, but I'd
rather just not use attributes there at all.