audio: convert audio samples to uniform format
When audio files are loaded, they are now converted to the same format
and sample rate used by the audio mixer.  This simplifies the mixer
code.  Since the sample rate conversions are done ahead of time, the
resampling code could be upgraded to a higher quality algorithm without
worrying about doing it in real-time.