Rip out asynchronous resource loading
This is a major change.  Previously, resources were loaded in various
background threads and using AIO callbacks.  However, this resulted in
complicated code and there was still the issue of how to signal that a
resource or set of resources is loaded.

The new decision is that general resource loading is outside the scope
of the project.